Script to intall programs in madagascar package
That SConstruct script is used to install a user C program in the madagascar
package. Put the program in a user directory inside madagascar users
directory with the 'M' preffix and '.c' suffix, example Mvfsa.c would be
the source code for a 'sfvfsa' madagascar program.

After that, put the name of your program (without 'M' and '.c') in the
'progs' variable in that SConstruct and run it in your user's directory at your
Madagascar installation to compile with 'scons' command.
Go to the top Madagascar directory and run 'scons install',
to install the compiled program in your Madagascar package.