diagrams and experiments directories
In diagrams directory put the diagrams developed with 'dia' in order to
explain and ilustrate the algorithms.

In experiments directory, put numerical experiments to test results of
the code in sevel circuntances and input parameters.
TDD directory has the purpose to test the program code, if it's
compiling and runing, and the expected output is correct for a normal
use. In the experiments directory, test numerical results and iputs, the
purpose it's not the code, but to improve the results