sfvfsacrenh to Madagascar package
Program 'sfvfsacrenh' gets CRS parameters RN, RNIP and BETA using Very
Fast Simulated Aneeling algorithm (VFSA). It tries to fit the
non-hyperbolic CRS approximation surface in the data cube A(m,h,t) using
a semblance criteria, if it fits, those parameters in the approximation
that produced the best semblance are the optimized ones.

Translate the original code documentation and functions names to
english. It's compiled and installed in Madagascar package 2.0, but is
missing the usage test.