Data cube interpolation to increase m sampling
CRE gather trajectory in the m,h plane has not well sampled values in m
for constant offset gathers. This trajectory is a curve that intercets
constant offset Gathers in an arbitrary m position. Use data interpolation
to increase CMP data sampling, in order to get traces (a m,h pair point)
closer to the correct m,h in the CRE trajectory that intersects a
h=constant in a constant offset Gather.

In order to do so, interleave zero traces in original data offset
Gathers and interpolate each section with adaptative preditive error
filters (PEF) method. Concatenate all interpolated section in a
resampled data cube (Double of traces in CMP coordinate, half of CMP
sampling, keep time and offset sampling same as the original data cube).

PEF coeficients calculation step parameters: Jump parameter, jumps every
group of 2 traces to faster the data interpolation. a is number of
coeficients in time and space, in that order. rect1 is the smothing
window in time, rect2 is the smothing window in space.

Interpolation step parameters: exact is to keep the original traces and
only interpolate the zero traces. mask receives a mask file to inform
wich traces to keep (original traces) and wich ones to interpolate (zero
ones).