Get CRE gather for any m0
Get a CRE gather for a given m0. That SConstruct generalize the
buildCREGather experiment wich m0=5Km only. In that new experiment
someone can choose a central CMP m0 and run the algorithm in order to
get the CRE Gather.

Observe that for a new m0 someone will have to give a new normal ray
traveltime t0 and calculate the CRS parameters, CRS trajectory and CRS
traveltime curve because they depend on t0 and RN, RNIP and BETA values.

One way is to give a value of t0 to the program. Another way is to
calculate the CRE gather for every t in a window in time axis of the
central trace (identified by m0 and h=0) in the zero ofsset section, and
get the major value of semblance, and the CRS parameters related to that
value.