Get CRE gather for several m0's and t0's
Save the CRS parameters obtained in the optimization step in order
to build the radiusgram, anglegram and semblancegram.

Radiusgram will be a section with distance in horizontal axis, time in
vertical axis and RNIP values obtained.

Anglegram will be a section with distance in horizontal axis, time in
vertical axis and BETA values obtained.

Semblacegram will use the same axis as the other sections, but with
semblance values calculated with the CRE traveltime curve on the CRE
Gather, is not the semblance value obtained in the optimization step,
because this is the value for the CRS surface, and we are interested in
semblance values in CRE domain.

The time axis is the t0 time and distance is the m0 position.