CRE stack and VFSA bug correction
Program Mcrestack.c stacks the cre Gathers, given cre traveltime curves
and cre Gather sections. It sums the amplitudes in the cre Gather
section through a cre traveltime curve. That sumation is done for all
t0's and m0's to form the cre stacked section.

VFSA was not converging because the data cube was not loaded properly,
to load only a section of the data cube was proved not a good practice,
and the effect in performance it's not significative.