Organize creGathers and creTimeCuves by t0 and m0
Concatenate the traveltime curves in a 3 dimension cube. Where n1 is the
index of the cre sample in the traveltime curve, n2 is the t0 coordinate
and n3 is the m0 coordinate. The n1 coordinate can be associated with
m,h pair by its respective coordinate vector produced with
sfgetcregather program in m= parameter.

Concatenate the cre Gathers in a 4 dimension hypercube. Where n1 is the
time, n2 is the cre gather index (see the explanation above), n3 is the
t0 coordinate and n4 is the m0 coordinate.