Correct bug in Mcrestack.c input dimensions
Correct the way that sfcrestack get input dimensions of traveltime
curves and cre Gather cube. The first dimension in traveltime curves
input is a index that identify the sample in cre domain, it corresponds
to the second dimension in the cre Gather painels in creGatherCube file.

So, to stack in the cre gather domain using a given traveltime curve,
someone doesn't need to input the coordinates, the program identify the
sample by index in the cre Gather domain, if you produce the input data
using this SConstruct flows.