Correct input geometry in Mcrestack.c
Program sfcrestack is now receiving the correct geometry of input files.
It receives the traveltime cre curve for each m0, t0 pair in the stacked
section and the cre Gathers.

Traveltime curve n1 is the sample of cre Gather, n2 is t0 and n3 is m0
coordinate. For cre Gathers, n1 is the time, n2 is the cre sample, n3 is
the t0, n4 is the cmp.