Reproducible papper images from modeling
Vpl images should be first converted for jpeg with jpegpen program and after they should be converted to pdf with img2pdf program, because the conversion directly from vpl to pdf with pdfpen doesn't provide a good resolution.

The program img2pdf can be installed with:

~$ sudo apt-get install img2pdf

SConstruct in the modeling directory will run the numerical experiment
build figures and converts all the images to pdf in order to build the
reproducible papper.