Define the CFLAGS during compilation
Alter the definifition of the CFLAGS to inform the preprocessor a
simbolic constant definition as TDD=1. It'll turn possible to build TDD
tests in the programs using a key variable to set if the the test is on
or not.

If the TDD variable is set, all the tests inside the programs will be
compiled. If it's not, the tests will be ignored by the compiler.