If to get tdd variable to compile with tdd option
Implements a if block to check if the tdd variable is set. Soemone can
set that passing 'scons tdd=1' in the compilation step of the programs
in this repository.

That option is to inform the preprocessor to compile the TDD code
with the normal code of the programing to allow TDD tests.