Resolve test on issue #3
Use the nonhyperbolic CRS traveltime equation in the building of the
stacking traveltime curve instead of the CRE equation. The idea is to
increase the region of convergence of the traveltime equation to improve
stacking results. It uses the CDS condition (RN=RNIP) in the traveltime
equation.

The same test was acomplished with the CDS equation, but the best
results were achieved using nonhyperbolic CRS with the CDS condition.