Modify file name of Mvfsacrenh.c to Mvfsacrsnh.c
The program name Mvfsacrsnh.c is better than Mvfsacrenh.c, because it
does not use CRE method. Depite it can be used as one step in the CRE
methodology, it uses non-hyperbolic CRS in order to obtain RN, RNIP and
BETA parameters.

This program (Mvfsacrsnh.c) was accepted in the Madagascar package. This
is anothe reason to change the program name.