Delete useless examples Resolve #34
Those examples are no longer required. They were generalized in the
SConstruct in the 'fullInterpolationAndStack' directories that represents
the complete CRE Stacking workflow.

The usage example of PEF interpolation remains because this part can be
used for general purposes and process. Modeling and Vfsa usage examples
were moved to Vfsa package repository