Use an aperture parameter in sfgetcregather #38
The 'aperture' parameter in 'Mgetcregather.c' defines the maximun
number of traces in a cre gather. This is used to avoid generation
of a big cre gather with many traces that won't be used in stacking.

It also avoid to take offset coordinates that have CMPs far away from
Central CMP m0 that may cause a bug. For a given offset h, the cmp
coordinate m could be outside of the datacube because the CRE trajectory
in m,h plane is a parabola and CMP coordinates increases fast in relation
to offset.