Stablish recipes as new dependecy madagascarRecipe #38
The 'creStack.py' script is divided in several modules of a new library
'madagascarRecipes' to stablish a dependency. Because this modules are
used in other packages, such as 'vfsa' or 'diffractionSimulator'.

With the modules in a package, this package can use one version of the
library without dependency conflict.

The library can be found in:
https://github.com/Dirack/madagascarRecipes