Error report function to madagascarRecipes #37
This script defines a function that reports a error message when build
fails. The user can control the message to be reported.

SConscript has changed to use that error report. If build fails it calls
'print_build_failures' function and shows the string in 'message' variable

Alias function defines aliases to call specific builds through command line.
example:

~$ scons -f SConscript model

It will call the building of the object defined to 'model' alias