Bug Report information to hotfix #38
Using aperture=49 in the the program sfgetcregather with i=101 and
j=73 runs without bug. Meanwhile, running with aperture=50 ends with
the bug referenced in this issue.

This probably occurs because trace number 50 has a CMP coordinate out
of the model, m[50]=5.026 (I've checked in 'creMhcoordinate' file).

Please reference this commit to hotfix/issue/38 discussion.