Aperture should be the same for sfcrestack and sfgetcregather #44
As referenced in bug report #49, aperture parameter should be the
same for the programs sfcrestack and sfgetcregather to avoid building
errors in the stacking. Aperture parameter is the number of offsets
to consider in the cre trajectory for sfgetcregather and the number
of offsets to stack in sfcrectack.

This parameter was introduced in sfgetcregather program to correct
bug #38