Correct variable 'file' way to receive input #44
The variable 'file' was loaded before cre stacking loop. It used glob
method to receive the stacked traces list available in the directory
to sort and build the stacked section.

The problem is if the variable is loaded before the loop, glob won't
find any stacked traces, because they are not build already! So, in order
to explicit tell scons to load 'files' variable after the loop, put the
variable as an empty list and load the values during the for loop with
append method