Use define LOGGING to allow warnings #55
Use define LOGGING macro to allow logging and warning option
during program compilation. Someone only have to comment or
uncomment the line with the macro definition to use or not the
warnings during the program execution.

This option is for degub purpose only. To verbose execution,
use verb parameter available for sfrnip2vrms