Refactoring of the travis CI tests #55
1. Conditional builds in travis CI

Configure .travis.yml to allow conditional builds of
the stages with environment variables evaluated as
true. Each job has its own name.

2. Include unit tests in the .travis.yml file

Now, travis will run 3 stages of tests, unit tests calling
make in the TDD directory; modeling step, calling the unittest
module for each test script in the TDD directory; and deploy
stage, that should be written.

This is done with test_travis_docker.sh script, that generate a
docker container with cregatherstack:1.0.0 image where Madagascar is
installed and pre configured. This script is called in the
.travis.yml file

3. New tests for negative values

The function calculateVrmsSectionForRnipVector returns 0
for negative values. It is added a test for this condition
and a test for a random set of rnip values in the test suite