[ADD] Deploy scripts
Scripts in the deploy directory builds template files with '.deploy'
suffix. Travis CI calls these scripts and deploys the generated files
in a draft release to be published.

The script 'build_release_message.sh' generates a release description
with relevant commits it gets from history. It uses marks in the
commit subject (described in the script header) to filter relevant
commits and group them in sections on the release description
based on this marks in a tag.deploy file.

The script 'build_changelog_update.sh' uses the file generated by the
previous script to update the CHANGELOG, it generates a template file
CHANGELOG.deploy with updated CHANGELOG.

Finally, the 'build_dockerfile.sh' generates the Dockerfile to be
used to package building of this package current version in a
template file, Dockerfile.deploy.