New experiment for the new package version #61
This experiment is a test for the new programs of the creGatherStack package
that uses vfsa 2.0 as a dependency. In this new version, the programs do the
process of cre stacking for all (t0,m0) pairs at once reading the parameters
cube generated by sfvfsacrsnh.

Run this test with the following command:

~$ scons nm0=2 nt0=2

For a quick test