Makefile control parameters #61
Use makefile to control the number of offsets to interpolate nhi
in the PEF interpolation, it will be equal to the value of the aperture
(offsets to stack) in CRE stacking.

Stacking aperture is arbitrarily choosen, theres no need to be equal to nhi,
but its maximum value is the number of offsets available in the seismic datacube.
Its minimum value is of course at least 1 offset to stack. So,

aperture belongs to [1,nhi]