Remove deprecated experiments and creGatherStack.py lib
Now, the 'creGatherStack.py' is being developed as a Madagascar recipe
to be installed in /home/rodolfo/madagascar-3.1.1/book/Recipes directory in Madagascar local installation.
This recipe is being developed in another github repository and it is now a creGatherStack package
dependency. This library is now available in https://github.com/Dirack/recipes repository.

Also, creGatherStack.py functions have changed, wich means that old usage examples in experiments
directory are now deprecated and replaced by new ones that use the current library version

And the CDS stacking option was abandoned also in order to simplify this
package usage and functionality, so this feature and programs related
are also deprecated